+++ internal/service/hearings/hearings.go
@@ -23,7 +23,8 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/brurbanko/mercury/internal/crawler"
+	"github.com/brurbanko/mercury/pkg/crawler"
+
 	"github.com/brurbanko/mercury/internal/database"
 	"github.com/brurbanko/mercury/internal/domain"
 	"github.com/rs/zerolog"
